Disable march=native for SSE2
Without this patch, the code would only build binaries with SSE2,
if the build machine CPU supported SSE2.

And through the -march=native flag, it would then use more CPU features
available on the build CPU that might not be available
on the target machine CPU. That can cause an illegal instruction
exception on the target machine.

This patch changes behaviour so that only configure options determine the
result of the build, regarding SSE2.

See https://reproducible-builds.org/ for why this is good.

This PR was done while working on reproducible builds for openSUSE.