tests: Generate encoded "lorem" files from a single source.
Using a single source for files with the same content just using
different encoding is always a good thing.  In this case it showed
that the UTF-16 and UTF-18 files could not be easily reproduced from
the latin1 encoded file:  These files include a BOM and are
big-endian.  When trying to recreate these files on my machine, I got
litte-endian files at first.