tests: Don't sort filenames for archives.
For some reason the files within the archies contained in the source
distribution have not been sorted. To ensure the archives created by
the script having the same order as the archives contained in the
source distribution, force "ls" no *not* sort the filenames, but use
the order on the command line.