Add Macports moar +pager variant (moar-pager)
In macports/macports-ports#21604, the MacPorts team merged a way of
installing the `moar` pager (support for `moar` was added to `ugrep`
@295382bc72) that renames the installed binary to `moar-pager` because
`moar` conflicts with name of the Rakudo virtual machine (MoarVM). This
is not the default variant, but it will be used by MacPorts users who
prefer `moar` to `less` and also develop nqp or Rakudo (and therefore
require MoarVM).

This patches `ugrep` to check the `$PAGER` command for `moar-pager`
before `moar`, the same as we now do in the `ugrep` port.

Merging this will allow us to no longer require a patch to the port
distribution and continue distributing upstream unmodified, which is our
preference.