Remove generated files from source tree
See
https://stackoverflow.com/questions/15745534/should-i-include-configure-and-makefile-in-a-github-repository

Also many other references why generated files should not be checked in.
Git does not preserve file timestamps, so users who clone will need to
re-generate these files, which would make their source tree "modified"
in git's view.

One valid use to keep these in SCM is to monitor change and verify
unexpected changes. However, this implies the maintainer is familiar
with automake's internal and thus able to discern valid change from
incorrect change.

If that is the goal, the proper way to implement it is to create a
'generated' branch or such which contains the gnerated files, but keep
master clean. Diff master to generated from time to time, or integrate
master into generated occasionaly/regularly/at-release-time and verify
that the generated branch contains only expected change.

If the intent was to keep an eye on change, it wasn't working. The build
setup has many problems that have gone unnoticed. I will fix them in the
following commits in as small sizes to make it easy to document /
understand and review.