Make CPU feature detection scalable, detect SSE2
With an eye on allowing CPU-specific optimizations to be added
in the future, and on making the produced binary portable to all
CPUs of the target instruction set, this commit refactors the HW
variables that stores CPU capabilities, and the way it's accessed.

The uint64_t variable HW is replaced with a const bitfield (struct
SIMD_caps).

The have_HW_*() accessor functions are removed.
The detection of known features is moved from accessors to the get_HW()
function, and the hardcoded bit positions (62, 37), which require
significant effort to decode by a human examining the code, are replaced
with a more clear and easier to read representation.

Removed this:
inline bool have_HW_AVX2()
{
  return HW & (1ULL << 37);
}

Replaced with this:
  caps.avx2      = !!(CPUInfo7[1] & bit_AVX2);     /* ebx_cpuid7,5 */

Additionally, detection for SSE2 is added in the same fashion as
AVX2/AVX512_BW.

A future commit will add NEON detection into the same detection
framework (bitfield struct computed at startup)

Implement Matcher::simd_match_sse2()

Without this function, a binary built with extensions had no way to
access non-SSE2 code. If running on a CPU not capable of SSE2, it would
just attempt to run SSE2 code and (probably) crash.

Fixes Genivia/ugrep#368

Changed ax_ext.m4 so it defines implementation FLAGS settings
SIMD_FLAGS_SSE2, SIMD_FLAGS_AVX2, etc, and also application FLAGS
settings (SIMD_FLAGS)

The implementation flags will be used to build libraries like
libreflex-sse2.a, libreflexx-avx2.a. etc which contain the respective
logic implementation using intrinsics. They are only built when the
compiler supports the respective instruction set and the corresponding
HAVE_{SSE2,AVX2...} define is set, such that libreflex-sse2.a is built
only with -msse2 and not with other flags, and similarly the -avx.a
counterpart is only built with -mavx2 and not other flags.

GCC manual states:

   These options enable GCC to use these extended instructions
   in generated code, even without -mfpmath=sse.  Applications
   that perform run-time CPU detection must compile separate
   files for each supported architecture, using the appropriate
   flags.  In particular, the file containing the CPU detection
   code should be compiled without these options.

The SIMD_FLAGS contains all the desired and compiler-supported
optimizations, and is meant to allow the compiler to automatically
vectorize for those architectures. This ability will not be used in
ugrep for generic compiles, but may be useful for users who want to
build the app targetting a specific CPU. IOW, distros that distribute
ugrep will compile without auto-vectorization, and the SSE2/AVX/NEON/etc
instruction will only be used if detected at runtime.