Fix "No rule to make libviiz.a"
When running make in the src folder, without building the lib first, the
following error was reported:

make: *** No rule to make target '../lzma/C/lib/libviiz.a', needed by
'ugrep'.  Stop.

This means ugrep's dependency chain is incomplete.
Getting the "No rule" error is the best possible manifestation.

In a build tree where aa "libviiz.a" file exists, but is outdated with
respect to its sources, it will be used to build ugrep (ie, an
out-of-date) lib is used because the file exists, but make doesn't have
the rules that would tell it how to check its freshness.

This is a pathological problem in projects which split convenience
libraries from the parts of the project where they are needed (typically
"lib" directory separate from "src"). This common scenario happened
here.

The solution implemented in this commit is to move the specifications
from 'lib/Makefile.am' into 'src/Makefile.am' and delete the unecessary
lib Makefile. This way, when building ugrep in the src directory, the
full dependency chain is available to make, and the binary will be
updated taking into consideration all updated sources. The stale build
problem cannot occur.

I found this problem while editing the process using the
Autotools Integration extension for vscode, which reports the broken
dependency chain as project diagnostics.