Trim ax_ext.m4 to remove unneeded capability checking
This software only cares about SSE2, AVX2 and AVX512_BW capabilities
currently. The other extensions are removed, to make the configuration
process quicker, and less unnecesarily verbose. There is no need to
check features that are not used.

In the future, when other extensions are supported, the GNU website
should be checked for an updated version of the script, and any needed
extensions copied from there. See the ax_ext.m4 header for the URL.