Moved the install-exec-hook where it belongs, and cleanup
The install hook is used to create hardlinks to the installed executable
(the ugrep+, ug and ug+ aliases are hardlinks to the main executable,
ugrep).

Cleanup means that instead of creating 4 copies of the same file, the
hook now attempts to hardlink, and only if that fails, a copy is made.
The copy is made using the $(INSTALL) program which is different for
every platfom, and contains platform specific know-how to install
programs. The old technique with 'cp -f' and 'rm -f' is not as portable
as $(INSTALL)

This hook belongs in the same Makefile where ugrep is built, not at the
root Makefile, so it's moved there.