Remove shell completion dependencies
1. bash-completion is not bash, but a third-party library of completions
for programs that did not provide their own completions. It is the wrong
thing to depend on. The correct thing is bash itself, and specifically
the documented behaviour that system-wide completions are expected by
bash to be in /etc/bash_completions.d.

There is not a single byte in the bash-completion package that is needed
to build ugrep, so there should not be a PKG_CHECK_MODULES about it.

2. Similarly, there is nothing in the fish package that is needed to
build ugrep

This commit removes both bash-completion and fish dependencies and uses
the well-known specification that is the directory names where the
completion files should be placed.

Case in point:

On Ubuntu, the fish completion is not installed. The bash completion is
installed in the wrong place (bash-completion's share directory), and
only accidentaly, because the Ubuntu build servers happen to have it
installed. Installing it in the wrong directory means that on systems
that have bash but not the 'bash-completion' library of orphan scripts,
the ugrep completion is not available, but should be.