Generate manpage from template
The canonical/proper/enlightened way to output manpages with autotools
is to store a template in the source repo, with placeholders for project
name and version, and substitute the proper values when the package is
built for installation.

This way the manpage can include effects of configuration, eg. if a user
disables a particular feature, it can be omitted from the installed
manpage.

In this way if the manpage does not change from a release to the next,
its header will indicate the last update time, and the current release
version. The "last updated" date ought to not change if the doc is not
changed, but the version identifier should correspond with the installed
binary's --version output.

This commit implements this canonical method.

The man/ugrep.1 file is replaced with man/ugrep.1.in template, and the
template is hydrated at build time with the project name and actual
released version number.

The ug.1 symlink is also generated the proper way, with
AC_CONFIG_LINKS([man/ug.1:man/ugrep.1]) rather than hacking it in though
a hook. A hook is not needed, because autoconf has built-in support for
aliased files.