Install _configured_ bash completions
The bash completion file completions/bash/ug is replaced with a
template, completions/bash/ug.in which contains placeholders for
PACKAGE_NAME and is hydrated at install time.

Completions for the aliases ugrep+, ug, ug+ are also installed
canonically using AC_CONFIG_LINKS in configure.ac. Previous
"install-data-hook" is removed, because the functionality for installing
symlinks is built into auto tools. It automatically does the right thing
of symlink/hardlink or copy depending on the destination OS's
capabilities. Ie, this is the portable way to install aliases.