Install _configured_ zsh completions
This commit removes the generated completion files from git, and
generates completions at build/install time using compgen.

The zsh completions are installed the autotools way, and the install
hook is removed, as it's unnecessary (and breaks distcheck)

CAUTION: The .DELETE_ON_ERROR GNU Make directive is required in
src/Makefile.am to delete the generated zsh completions in case
generation fails. This is a GNU-specific feature, which makes the build
somewhat unportable. It will still build on non-GNU make, but if
completion generation fails, the incomplete produced completion file
will be installed, if "make install" is run again after the first
failure without fixing the error.

The compgen file uses bash syntax ($'...') so I changed the shebang to
require bash instead of sh.

The completion rule is moved to the src directory because it depends on
src/ugrep binary, so it belongs in the same Make scope.