Implement runtime NEON detection
The home-grown NEON detection in configure.ac was breaking cross
compiling because relied on -march=native to detect if the compiler was
running on a NEON capable machine, which would never be true when
cross-compiling.

Added compiler-capability detection for NEON to ax_ext.m4, to detect if
arm cross compilers support -mfpu=neon.

Added runtime test for NEON capabilities for Arm-7a CPUs. Arm8 always
has NEON, and ARMv6 never has NEON, and expanded the conditional
compilation to build a NEON-specific implementation of the matcher if
the compiler supports it.