Remove homegrown PCRE2/Boost.Regex detection
Autotools has built a in mechanism for finding packaged libraries:

PKG_CHECK_MODULES

Boost.Regex does not distribute a pkg-config file, apparently because
the name of the library is inconsistent across distros, so it is
detected with AC_SEARCH_LIBS and AC_CHECK_HEADER

Deleted m4/ax_check_pcre2.m4 and m4/ax_boost_regex.m4. Custom scripts
are not only unnecessary, but these were broken. Not supporting
cross-compilation, and relying on hardcoded paths means they wouldn't
work on all distros.

The boost libs in particular, are complex beasts, and an attempt by a
third party to provide configuration that the authors themselves don't
provide, is error prone, and the deleted script was riddled with errors.
On my Ubuntu-16 system, the library is not boost_regex-mt, but simply
boost-regex to name only a more glaring problem. It is unlikely that a
third party would be able to provide a portable way to detect libboost,
so this script should have never existed.

Also, the deleted scripts did not provide an error to the user at
configuration time if both PCRE2 and boost-regex were not found; This
most elementary failure to report missing dependencies is what told me
that these hacks must be removed.