Implement correct version number string
The PACKAGE_VERSION define is used to set UGREP_VERSION.

1. There is no need to edit ugrep.hpp for every release.

2. In an autotools project, the source-of-truth as to the program
   version is the version argument to the AC_INIT macro in configure.ac

Case in point, the configure.ac's version string had fallen behind to
4.5 while ugrep.hpp indicates 5.0.0

Other software like the VSCode Autotools Integration understand a
project's version to be that defined in the AC_INIT macro; it is a
standard, known interface to obtain an arbitrary project's basic
coordinates like name, version, homepage and bugreport contact.

Also added the github Issues page and ugrep.com as the bugreport contact
and homepage info, respectively.