Adapt test suite to automake Parallel Harness
- the two tests, single and multi-threaded tests can now run
  simultaneously (they use the "st" and "mt" subdirectories as
  temp dirs so they do not conflict)

- The verify.sh is adapted to run the directory tests in the
  $builddir/src CWD, and the file tests in the $srcdir/tests CWD

This setup is not much cleaner than before. These tests will be a real
challenge to debug if necessary.

Adapting them to the Parallel Harness means they can run as standard
"make check" and not the home-grown "make test". This also means they
will automatically run during "make distcheck", which is the main
motivation for this adaptation.