Remove built windows binaries from git
The upstream builds binaries for windows do not belong in SCM.

They were previously included in the repo with the reasoning that they
are hard to reproduce. Even if that claim is true, they are generated
files and don't belong with source files. What if OS/2 binaries were
hard to reproduce, or macos binaries, or Win32/ARM or others? Should
they be stored in the source repo? No.

These were 64-bit and 32-bit binaries for X86 processors, but Windows
for ARM binaries were not included. Supposedly they are hard to
reproduce too, yet a double-standard is visible, as they are not deemed
by the author relevant enough to treat the same as x86 binaries.

The previous commits fix cross-compiling, which allows windows binaries
to be easily cross-compiled on linux:
configure \
  PKG_CONFIG_SYSROOT_DIR=/mnt/build/win32-x64-sysroot \
  PKG_CONFIG_PATH=/mnt/build/win32-x64-sysroot/usr/lib/pkgconfig \
  CXX=x86_64-w64-mingw32-g++-posix \
  --host=x86_64-w64-mingw32

As it currently stands, the project has other bugs that prevent windows
compilation on Linux using the mingw32 toolchain, but those bugs are not
related to the issue of using the source repository as a built binary
archive, including:

  -lwsock32 -lmsvcr110 are needed to link
  - ENABLE_VIRTUAL_TERMINAL_INPUT is a Windows SDK definition, which
    should be added as a dependency to configure.ac. Without it, the
    implemented VKey logic is incomplete - it fails to link with
    VKey::poll() unresolved symbol.
  - CRT expects to find a WinMain (ie, -mconsole compiler flag is
    missing)

So while the argument 'hard to reproduce' was never a valid reason to
store them in SCM, as of the autotools overhaul that fixes
cross-compiling, it's not even a true claim.

Fixes Genivia/ugrep#123